+++ tests/integration/commands/agilityArenaRecolor.test.ts
